fix integer overflow on issue #10

+++ assembly/nmd_x86_formatter.c
@@ -109,7 +109,7 @@ NMD_ASSEMBLY_API void _nmd_append_relative_address16_32(_nmd_string_info* const 
 	if (si->runtime_address == NMD_X86_INVALID_RUNTIME_ADDRESS)
 	{
 		/* *si->buffer++ = '$'; */
-		_nmd_append_signed_number(si, (int64_t)((int32_t)(si->instruction->immediate) + (int32_t)(si->instruction->length)), true);
+		_nmd_append_signed_number(si, (int64_t)(si->instruction->immediate + si->instruction->length), true);
 	}
 	else
     {
